Added Ko-fi Support
According to all known forms of aviation, there is no way a bee should be able to fly. Its wings are far too small to get its fat little body off the ground. The bee of course, flies anyway, this is because it doesnt care what humans think are possible.